Fixed dom parsing.
